requirements.txt
requests
PrettyTable
python-dateutil
humanize
marshmallow
